+++ modules/aft-account-provisioning-framework/versions.tf
@@ -7,7 +7,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 4.9.0"
+      version = ">= 4.27.0, < 5.0.0"
     }
   }
 }
