+++ modules/aft-iam-roles/admin-role/main.tf
@@ -5,7 +5,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 4.27.0"
+      version = ">= 4.27.0, < 5.0.0"
     }
   }
 }
