+++ modules/aft-iam-roles/service-role/main.tf
@@ -6,7 +6,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 4.27.0"
+      version = ">= 4.27.0, < 5.0.0"
     }
   }
 }
