+++ sources/aft-lambda-layer/aft_common/aft_utils.py
@@ -2,28 +2,30 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 import logging
-import os
+import random
+import time
+from functools import wraps
 from typing import (
     IO,
     TYPE_CHECKING,
     Any,
+    Callable,
     Dict,
     Iterable,
-    List,
     Optional,
     Sequence,
+    Tuple,
     Union,
-    cast,
 )
 
 from boto3.session import Session
+from botocore.config import Config
+from botocore.exceptions import ClientError
 from botocore.response import StreamingBody
 
 if TYPE_CHECKING:
     from mypy_boto3_lambda import LambdaClient
     from mypy_boto3_lambda.type_defs import InvocationResponseTypeDef
-    from mypy_boto3_organizations import OrganizationsClient
-    from mypy_boto3_servicecatalog import ServiceCatalogClient
     from mypy_boto3_stepfunctions import SFNClient
     from mypy_boto3_stepfunctions.type_defs import StartExecutionOutputTypeDef
     from mypy_boto3_sts import STSClient
@@ -36,140 +38,84 @@ else:
     StartExecutionOutputTypeDef = object
     STSClient = object
 
-SSM_PARAM_AFT_DDB_META_TABLE = "/aft/resources/ddb/aft-request-metadata-table-name"
-SSM_PARAM_AFT_SESSION_NAME = "/aft/resources/iam/aft-session-name"
-SSM_PARAM_AFT_ADMIN_ROLE = "/aft/resources/iam/aft-administrator-role-name"
-SSM_PARAM_AFT_EXEC_ROLE = "/aft/resources/iam/aft-execution-role-name"
-SSM_PARAM_SC_PRODUCT_NAME = "/aft/resources/sc/account-factory-product-name"
-SSM_PARAM_SNS_TOPIC_ARN = "/aft/account/aft-management/sns/topic-arn"
-SSM_PARAM_SNS_FAILURE_TOPIC_ARN = "/aft/account/aft-management/sns/failure-topic-arn"
-SSM_PARAM_ACCOUNT_REQUEST_QUEUE = "/aft/resources/sqs/aft-request-queue-name"
-SSM_PARAM_AFT_ACCOUNT_PROVISIONING_FRAMEWORK_LAMBDA = (
-    "/aft/resources/lambda/aft-invoke-aft-account-provisioning-framework"
-)
-SSM_PARAM_AFT_CLEANUP_RESOURCES_LAMBDA = "/aft/resources/lambda/aft-cleanup-resources"
-SSM_PARAM_AFT_EVENTS_TABLE = "/aft/resources/ddb/aft-controltower-events-table-name"
-SSM_PARAM_AFT_SFN_NAME = (
-    "/aft/account/aft-management/sfn/aft-account-provisioning-framework-sfn-name"
-)
-SSM_PARAM_AFT_DDB_REQ_TABLE = "/aft/resources/ddb/aft-request-table-name"
-SSM_PARAM_AFT_DDB_AUDIT_TABLE = "/aft/resources/ddb/aft-request-audit-table-name"
-SSM_PARAM_AFT_REQUEST_ACTION_TRIGGER_FUNCTION_ARN = (
-    "/aft/resources/lambda/aft-account-request-action-trigger-function-arn"
-)
-SSM_PARAM_AFT_ACCOUNT_REQUEST_AUDIT_TRIGGER_FUNCTION_ARN = (
-    "/aft/resources/lambda/aft-account-request-audit-trigger-function-arn"
-)
-SSM_PARAM_AFT_ACCOUNT_REQUEST_PROCESSOR_FUNCTION_ARN = (
-    "/aft/resources/lambda/aft-account-request-processor-function-arn"
-)
-SSM_PARAM_AFT_CONTROLTOWER_EVENT_LOGGER_FUNCTION_ARN = (
-    "/aft/resources/lambda/aft-controltower-event-logger-function-arn"
-)
-SSM_PARAM_AFT_INVOKE_AFT_ACCOUNT_PROVISIONING_FRAMEWORK_FUNCTION_ARN = (
-    "/aft/resources/lambda/aft-invoke-aft-account-provisioning-framework-function-arn"
-)
-SSM_PARAM_AFT_ACCOUNT_PROVISIONING_FRAMEWORK_CREATE_ROLE_FUNCTION_ARN = (
-    "/aft/resources/lambda/aft-account-provisioning-framework-create-role-function-arn"
-)
-SSM_PARAM_AFT_ACCOUNT_PROVISIONING_FRAMEWORK_TAG_ACCOUNT_FUNCTION_ARN = (
-    "/aft/resources/lambda/aft-account-provisioning-framework-tag-account-function-arn"
-)
-SSM_PARAM_AFT_ACCOUNT_PROVISIONING_FRAMEWORK_PERSIST_METADATA_FUNCTION_ARN = "/aft/resources/lambda/aft-account-provisioning-framework-persist-metadata-function-arn"
-SSM_PARAM_AFT_ACCOUNT_PROVISIONING_FRAMEWORK_NOTIFY_ERROR_FUNCTION_ARN = (
-    "/aft/resources/lambda/aft-account-provisioning-framework-notify-error-function-arn"
-)
-SSM_PARAM_AFT_ACCOUNT_PROVISIONING_FRAMEWORK_NOTIFY_SUCCESS_FUNCTION_ARN = "/aft/resources/lambda/aft-account-provisioning-framework-notify-success-function-arn"
-SSM_PARAM_AFT_MAXIMUM_CONCURRENT_CUSTOMIZATIONS = (
-    "/aft/config/customizations/maximum_concurrent_customizations"
-)
-SSM_PARAM_FEATURE_CLOUDTRAIL_DATA_EVENTS_ENABLED = (
-    "/aft/config/feature/cloudtrail-data-events-enabled"
-)
-SSM_PARAM_FEATURE_ENTERPRISE_SUPPORT_ENABLED = (
-    "/aft/config/feature/enterprise-support-enabled"
-)
-SSM_PARAM_FEATURE_DEFAULT_VPCS_ENABLED = (
-    "/aft/config/feature/delete-default-vpcs-enabled"
-)
-SSM_PARAM_ACCOUNT_CT_MANAGEMENT_ACCOUNT_ID = "/aft/account/ct-management/account-id"
-SSM_PARAM_ACCOUNT_AUDIT_ACCOUNT_ID = "/aft/account/audit/account-id"
-SSM_PARAM_ACCOUNT_LOG_ARCHIVE_ACCOUNT_ID = "/aft/account/log-archive/account-id"
-SSM_PARAM_ACCOUNT_AFT_MANAGEMENT_ACCOUNT_ID = "/aft/account/aft-management/account-id"
-
-SSM_PARAM_ACCOUNT_AFT_VERSION = "/aft/config/aft/version"
-SSM_PARAM_ACCOUNT_TERRAFORM_VERSION = "/aft/config/terraform/version"
-
-SSM_PARAM_AFT_METRICS_REPORTING = "/aft/config/metrics-reporting"
-SSM_PARAM_AFT_METRICS_REPORTING_UUID = "/aft/config/metrics-reporting-uuid"
-
 logger = logging.getLogger("aft")
+
+BOTO3_CLIENT_ERROR_THROTTLING_CODES = [
+    "ThrottlingException",
+    "TooManyRequestsException",
+    "RequestLimitExceeded",
+]
+
+
+def resubmit_request_on_boto_throttle(
+    func: Callable[..., Any],
+    max_requests: int = 3,
+    max_sleep_sec: int = 16,
+) -> Callable[..., Any]:
+    """
+    Decorator to automatically resubmit boto3-based API calls on throttling errors.
+
+    This decorator will re-submit up to max_requests INITIAL requests, NOT
+    accounting for built-in boto3 retries. Each INITIAL request is independent
+    of the previous requests. This may result in duplicate work being performed.
+    As such, this decorator should ONLY be used with read calls or
+    write calls that are idempotent.
+
+    For example, using the default boto3 config of 3 retries (4 attempts) with
+    max_requests = 3 will result in 12 total calls made to the AWS service.
+    """
+
+    @wraps(func)
+    def wrapper(*args: Optional[Tuple[Any]], **kwargs: Optional[Dict[str, Any]]) -> Any:
+        jitter = float(
+            f"{random.random():.3f}"  # nosec B311: Not using random numbers in a security context
+        )
+        retry_sleep_sec = min(2 + jitter, max_sleep_sec)
+
+        requests = 0
+        while True:
+            try:
+                return func(*args, **kwargs)
+            except ClientError as e:
+                if e.response["Error"]["Code"] in BOTO3_CLIENT_ERROR_THROTTLING_CODES:
+                    if requests >= max_requests:
+                        logger.info(
+                            f"Exceeded max fresh-request retry attempts ({max_requests})"
+                        )
+                        raise e
+
+                    logger.info(
+                        f"Exceeded max boto3 retries on previous request. Retrying with fresh request in {retry_sleep_sec} seconds."
+                    )
+                    requests += 1
+                    time.sleep(retry_sleep_sec)
+
+                    # Clipped exponential backoff with 0-1sec random jitter
+                    retry_sleep_sec = retry_sleep_sec * 2 + jitter
+                    retry_sleep_sec = min(retry_sleep_sec, max_sleep_sec)
+
+                else:
+                    raise e
+
+            except Exception as e:
+                # Raise on all other exceptions
+                raise e
+
+    return wrapper
+
+
+def get_high_retry_botoconfig() -> Config:
+    # https://boto3.amazonaws.com/v1/documentation/api/latest/guide/configuration.html#using-the-config-object
+    return Config(
+        retries={
+            "total_max_attempts": 16,  # original request + 15 retries
+            "mode": "adaptive",  # "Retries with additional client side throttling"
+        }
+    )
 
 
 def emails_are_equal(first_email: str, second_email: str) -> bool:
     return first_email.lower() == second_email.lower()
-
-
-def get_ssm_parameter_value(session: Session, param: str, decrypt: bool = False) -> str:
-    client = session.client("ssm")
-    logger.info("Getting SSM Parameter " + param)
-
-    response = client.get_parameter(Name=param, WithDecryption=decrypt)
-
-    param_value: str = response["Parameter"]["Value"]
-    return param_value
-
-
-def get_ct_product_id(session: Session, ct_management_session: Session) -> str:
-    client: ServiceCatalogClient = ct_management_session.client("servicecatalog")
-    sc_product_name = get_ssm_parameter_value(session, SSM_PARAM_SC_PRODUCT_NAME)
-    logger.info("Getting product ID for " + sc_product_name)
-
-    response = client.describe_product_as_admin(Name=sc_product_name)
-    product_id: str = response["ProductViewDetail"]["ProductViewSummary"]["ProductId"]
-    logger.info(product_id)
-    return product_id
-
-
-def get_ct_provisioning_artifact_id(
-    session: Session, ct_management_session: Session
-) -> str:
-    client: ServiceCatalogClient = ct_management_session.client("servicecatalog")
-    sc_product_name = get_ssm_parameter_value(session, SSM_PARAM_SC_PRODUCT_NAME)
-    logger.info("Getting provisioning artifact ID for " + sc_product_name)
-
-    response = client.describe_product_as_admin(Name=sc_product_name)
-    provisioning_artifacts = response["ProvisioningArtifactSummaries"]
-    for pa in provisioning_artifacts:
-        if ct_provisioning_artifact_is_active(session, ct_management_session, pa["Id"]):
-            pa_id: str = pa["Id"]
-            logger.info("Using provisioning artifact ID: " + pa_id)
-            return pa_id
-
-    raise Exception("No Provisioning Artifact ID found")
-
-
-def ct_provisioning_artifact_is_active(
-    session: Session, ct_management_session: Session, artifact_id: str
-) -> bool:
-    client: ServiceCatalogClient = ct_management_session.client("servicecatalog")
-    sc_product_name = get_ssm_parameter_value(session, SSM_PARAM_SC_PRODUCT_NAME)
-    logger.info("Checking provisioning artifact ID " + artifact_id)
-    try:
-        response = client.describe_provisioning_artifact(
-            ProductName=sc_product_name, ProvisioningArtifactId=artifact_id
-        )
-        provisioning_artifact = response["ProvisioningArtifactDetail"]
-    except client.exceptions.ResourceNotFoundException:
-        logger.info("Provisioning artifact id: " + artifact_id + " does not exist")
-        return False
-
-    if provisioning_artifact["Active"]:
-        logger.info(provisioning_artifact["Id"] + " is active")
-        return True
-    else:
-        logger.info(provisioning_artifact["Id"] + " is NOT active")
-        return False
 
 
 def invoke_lambda(
@@ -221,72 +167,6 @@ def is_aft_supported_controltower_event(event: Dict[str, Any]) -> bool:
             return True
 
     return False
-
-
-def get_all_aft_account_ids(aft_management_session: Session) -> List[str]:
-    table_name = get_ssm_parameter_value(
-        aft_management_session, SSM_PARAM_AFT_DDB_META_TABLE
-    )
-    dynamodb = aft_management_session.resource("dynamodb")
-    table = dynamodb.Table(table_name)
-    logger.info("Scanning DynamoDB table: " + table_name)
-
-    items: List[Dict[str, Any]] = []
-    response = table.scan(ProjectionExpression="id", ConsistentRead=True)
-    items.extend(response["Items"])
-    while "LastEvaluatedKey" in response:
-        logger.debug(
-            "Paginated response found, continuing at {}".format(
-                response["LastEvaluatedKey"]
-            )
-        )
-        response = table.scan(ExclusiveStartKey=response["LastEvaluatedKey"])
-        items.extend(response["Items"])
-
-    aft_account_ids = [item["id"] for item in items]
-
-    if not aft_account_ids:
-        raise Exception("No accounts found in the Account Metadata table")
-
-    return aft_account_ids
-
-
-def get_accounts_by_tags(
-    aft_mgmt_session: Session, ct_mgmt_session: Session, tags: List[Dict[str, str]]
-) -> Optional[List[str]]:
-    logger.info("Getting Account with tags - " + str(tags))
-    # Get all AFT Managed Accounts
-    all_accounts = get_all_aft_account_ids(aft_mgmt_session)
-    matched_accounts = []
-    client: OrganizationsClient = ct_mgmt_session.client("organizations")
-    # Loop through AFT accounts, requesting tags
-    if all_accounts is None:
-        return None
-
-    for a in all_accounts:
-        account_tags = {}
-        response = client.list_tags_for_resource(ResourceId=a)
-        # Format tags as a dictionary rather than a list
-        for t in response["Tags"]:
-            account_tags[t["Key"]] = t["Value"]
-        logger.info("Account tags for " + a + ": " + str(account_tags))
-        counter = 0
-        # Loop through tag filter. Append account to matched_accounts if all tags in filter match/ are present
-        for x in tags:
-            for k, v in x.items():
-                if k in account_tags.keys():
-                    if account_tags[k] == v:
-                        counter += 1
-                        if counter == len(tags):
-                            logger.info(
-                                "Account " + a + " MATCHED with tags " + str(tags)
-                            )
-                            matched_accounts.append(a)
-    logger.info(matched_accounts)
-    if len(matched_accounts) > 0:
-        return matched_accounts
-    else:
-        return None
 
 
 def get_session_info(session: Session) -> Dict[str, str]:
