+++ sources/aft-lambda-layer/aft_common/notifications.py
@@ -4,10 +4,8 @@
 import logging
 from typing import TYPE_CHECKING
 
-from aft_common.aft_utils import (
-    SSM_PARAM_SNS_FAILURE_TOPIC_ARN,
-    get_ssm_parameter_value,
-)
+from aft_common.constants import SSM_PARAM_SNS_FAILURE_TOPIC_ARN
+from aft_common.ssm import get_ssm_parameter_value
 from boto3.session import Session
 
 if TYPE_CHECKING:
