+++ sources/aft-lambda-layer/aft_common/shared_account.py
@@ -4,17 +4,17 @@
 import logging
 from typing import Any, Dict, List
 
-from aft_common import aft_utils as utils
 from aft_common import ddb
 from aft_common.account_provisioning_framework import ProvisionRoles
-from aft_common.aft_utils import (
+from aft_common.aft_utils import emails_are_equal, get_high_retry_botoconfig
+from aft_common.auth import AuthClient
+from aft_common.constants import (
     SSM_PARAM_ACCOUNT_AUDIT_ACCOUNT_ID,
     SSM_PARAM_ACCOUNT_CT_MANAGEMENT_ACCOUNT_ID,
     SSM_PARAM_ACCOUNT_LOG_ARCHIVE_ACCOUNT_ID,
-    emails_are_equal,
 )
-from aft_common.auth import AuthClient
 from aft_common.organizations import OrganizationsAgent
+from aft_common.ssm import get_ssm_parameter_value
 from boto3.session import Session
 
 logger = logging.getLogger("aft")
@@ -33,7 +33,9 @@ def shared_account_request(event_record: Dict[str, Any], auth: AuthClient) -> bo
     ct_management_session = auth.get_ct_management_session(
         role_name=ProvisionRoles.SERVICE_ROLE_NAME
     )
-    orgs_client = ct_management_session.client("organizations")
+    orgs_client = ct_management_session.client(
+        "organizations", config=get_high_retry_botoconfig()
+    )
     for shared_account_id in shared_account_ids:
         response = orgs_client.describe_account(AccountId=shared_account_id)
         if (
@@ -71,6 +73,6 @@ def get_shared_ids(aft_management_session: Session) -> List[str]:
         SSM_PARAM_ACCOUNT_CT_MANAGEMENT_ACCOUNT_ID,
     ]
     return [
-        utils.get_ssm_parameter_value(session=aft_management_session, param=ssm_param)
+        get_ssm_parameter_value(session=aft_management_session, param=ssm_param)
         for ssm_param in shared_account_ssm_params
     ]
