+++ src/aft_lambda/aft_account_request_framework/aft_account_request_audit_trigger.py
@@ -6,8 +6,8 @@ import logging
 import sys
 from typing import TYPE_CHECKING, Any, Dict
 
-from aft_common import aft_utils as utils
-from aft_common import notifications
+from aft_common import constants as utils
+from aft_common import notifications, ssm
 from aft_common.account_request_framework import put_audit_record
 from aft_common.logger import configure_aft_logger
 from boto3.session import Session
@@ -31,7 +31,7 @@ def lambda_handler(event: Dict[str, Any], context: LambdaContext) -> None:
             if "eventSource" in event_record:
                 if event_record["eventSource"] == "aws:dynamodb":
                     logger.info("DynamoDB Event Record Received")
-                    table_name = utils.get_ssm_parameter_value(
+                    table_name = ssm.get_ssm_parameter_value(
                         aft_management_session, utils.SSM_PARAM_AFT_DDB_AUDIT_TABLE
                     )
                     event_name = event_record["eventName"]
