+++ src/aft_lambda/aft_account_request_framework/aft_cleanup_resources.py
@@ -6,11 +6,12 @@ import logging
 from typing import TYPE_CHECKING, Any, Dict
 
 from aft_common import codepipeline, ddb
-from aft_common.aft_utils import SSM_PARAM_AFT_DDB_META_TABLE, get_ssm_parameter_value
 from aft_common.auth import AuthClient
+from aft_common.constants import SSM_PARAM_AFT_DDB_META_TABLE
 from aft_common.logger import configure_aft_logger
 from aft_common.notifications import send_lambda_failure_sns_message
 from aft_common.organizations import OrganizationsAgent
+from aft_common.ssm import get_ssm_parameter_value
 
 if TYPE_CHECKING:
     from aws_lambda_powertools.utilities.typing import LambdaContext
