+++ src/aft_lambda/aft_customizations/aft_customizations_execute_pipeline.py
@@ -5,7 +5,8 @@ import inspect
 import logging
 from typing import TYPE_CHECKING, Any, Dict
 
-from aft_common import aft_utils as utils
+import aft_common.ssm
+from aft_common import constants as utils
 from aft_common import notifications
 from aft_common.codepipeline import execute_pipeline
 from aft_common.logger import configure_aft_logger, customization_request_logger
@@ -24,7 +25,7 @@ def lambda_handler(event: Dict[str, Any], context: LambdaContext) -> Dict[str, A
     session = Session()
     try:
         maximum_concurrent_pipelines = int(
-            utils.get_ssm_parameter_value(
+            aft_common.ssm.get_ssm_parameter_value(
                 session, utils.SSM_PARAM_AFT_MAXIMUM_CONCURRENT_CUSTOMIZATIONS
             )
         )
