+++ src/aft_lambda/aft_feature_options/aft_delete_default_vpc.py
@@ -4,8 +4,9 @@
 import inspect
 from typing import TYPE_CHECKING, Any, Dict
 
+import aft_common.ssm
 import boto3
-from aft_common import aft_utils as utils
+from aft_common import constants as utils
 from aft_common import notifications
 from aft_common.account_provisioning_framework import ProvisionRoles
 from aft_common.auth import AuthClient
@@ -53,7 +54,7 @@ def lambda_handler(event: Dict[str, Any], context: LambdaContext) -> None:
         regions = get_aws_regions(client)
 
         if (
-            utils.get_ssm_parameter_value(
+            aft_common.ssm.get_ssm_parameter_value(
                 aft_session, utils.SSM_PARAM_FEATURE_DEFAULT_VPCS_ENABLED
             ).lower()
             == "true"
