+++ src/aft_lambda/aft_feature_options/aft_enable_cloudtrail.py
@@ -4,7 +4,8 @@
 import inspect
 from typing import TYPE_CHECKING, Any, Dict
 
-from aft_common import aft_utils as utils
+import aft_common.ssm
+from aft_common import constants as utils
 from aft_common import notifications
 from aft_common.account_provisioning_framework import ProvisionRoles
 from aft_common.auth import AuthClient
@@ -48,14 +49,14 @@ def lambda_handler(event: Dict[str, Any], context: LambdaContext) -> None:
         )
 
         # Get SSM Parameters
-        cloudtrail_enabled = utils.get_ssm_parameter_value(
+        cloudtrail_enabled = aft_common.ssm.get_ssm_parameter_value(
             aft_session, utils.SSM_PARAM_FEATURE_CLOUDTRAIL_DATA_EVENTS_ENABLED
         )
-        s3_log_bucket_arn = utils.get_ssm_parameter_value(
+        s3_log_bucket_arn = aft_common.ssm.get_ssm_parameter_value(
             aft_session, "/aft/account/log-archive/log_bucket_arn"
         )
         s3_bucket_name = s3_log_bucket_arn.split(":::")[1]
-        kms_key_arn = utils.get_ssm_parameter_value(
+        kms_key_arn = aft_common.ssm.get_ssm_parameter_value(
             aft_session, "/aft/account/log-archive/kms_key_arn"
         )
         log_bucket_arns = get_log_bucket_arns(log_archive_session)
